+++ src/main/scala/com/dataintuitive/viash/functionality/arguments/FileArgument.scala
@@ -15,11 +15,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.functionality.arguments
+package io.viash.functionality.arguments
 
 import java.nio.file.Path
-import com.dataintuitive.viash.helpers.Circe.OneOrMore
-import com.dataintuitive.viash.helpers.description
+import io.viash.helpers.Circe.OneOrMore
+import io.viash.helpers.description
 
 @description("")
 case class FileArgument(