+++ src/main/scala/com/dataintuitive/viash/platforms/NativePlatform.scala
@@ -15,16 +15,16 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.platforms
+package io.viash.platforms
 
-import com.dataintuitive.viash.config.Config
-import com.dataintuitive.viash.functionality.Functionality
-import com.dataintuitive.viash.functionality.resources._
-import com.dataintuitive.viash.platforms.requirements._
-import com.dataintuitive.viash.config.Version
-import com.dataintuitive.viash.wrapper.BashWrapper
-import com.dataintuitive.viash.helpers.description
-import com.dataintuitive.viash.helpers.example
+import io.viash.config.Config
+import io.viash.functionality.Functionality
+import io.viash.functionality.resources._
+import io.viash.platforms.requirements._
+import io.viash.config.Version
+import io.viash.wrapper.BashWrapper
+import io.viash.helpers.description
+import io.viash.helpers.example
 
 @description("""Running a Viash component on a native platform means that the script will be executed in your current environment.
                |Any dependencies are assumed to have been installed by the user, so the native platform is meant for developers (who know what they’re doing) or for simple bash scripts (which have no extra dependencies).