+++ src/test/scala/com/dataintuitive/viash/MainNSTestNativeSuite.scala
@@ -1,6 +1,6 @@
-package com.dataintuitive.viash
+package io.viash
 
-import com.dataintuitive.viash.helpers.IO
+import io.viash.helpers.IO
 import org.scalatest.{BeforeAndAfterAll, FunSuite}
 
 import java.nio.charset.StandardCharsets